+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## master
+
+### Fixed
+
+- `.as_dyn_error()` method hygiene inside `Error` derive expansion.
+  ([#527](https://github.com/JelteF/derive_more/pull/527))
+
 ## 2.1.0 - 2025-12-02
 
 ### Added